Additional features added.
Added support for multiple signatures (up to 5) for PDF signing.
Added support for page selection (last page / first page).

When there is a signature in a certain place the option will not show,
meaning that the document can't be signed twice on same position.

The content model was extended to meet the requirements and also new
evaluator was added.